simulation/pyproject.toml
[project]
name = "oasis-simulation"
version = "0.2.0"
description = "Unified simulation framework for Oasis firmware with MCU emulation and MCP interface"
authors = [{ name = "Oasis Team" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]
fmi = [
    "fmpy>=0.3",
]
pdf = [
    "pymupdf>=1.23",
    "pdfplumber>=0.10",
]
all = [
    "pymupdf>=1.23",
    "pdfplumber>=0.10",
    "fmpy>=0.3",
]

[project.scripts]
oasis-sim = "mcp_server:run_stdio_server"
oasis-datasheet = "datasheet_ingestion.cli:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["behavioral*", "emulators*", "datasheet_ingestion*", "components*"]
